requirements.txt
asgiref==3.7.2
boto3==1.28.57
botocore==1.31.57
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coverage==7.3.0
cryptography==41.0.3
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.20
django-allauth==0.54.0
django-countries==7.5.1
django-crispy-forms==1.14.0
django-jasmine==0.5.1
django-storages==1.14.1
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.0
psycopg2==2.9.8
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redgreenunittest==0.1.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
stripe==6.6.0
typing_extensions==4.7.1
urllib3==1.26.16
